Entities setup
- Created models, migrations, factories and seeders
- Updated dbdiagram. It's much simpler than I made id
- Created controlles, but not any function yet
- Created a very basic routing and folder structure.
- Updated css a bit.

The server calls the welcome.blade.php file for every single path (we set this up).

The welcome.blade.php file calls app.ts though vite (in the head, together with the style).

App.ts uses the router and mounts app.vue on the #app div.

So now #app in welcome displays App.vue.

App.vue uses the router to serve the components.

The css file is updated to color code the stuff based on provenience.